+++ recon/named/gatt_prepare_write_rsp.c
@@ -1,3 +1,4 @@
+#include <stddef.h>
 /* readable reconstruction; identity: FUN_0005c310 @ 0x0005c310
  * public-name: gatt_prepare_write_rsp
  * durable-map: recon/catalogs/function_names_app.json
@@ -21,7 +22,7 @@
 #include <stdint.h>
 extern void log_msg_create_3arg(uint32_t, uint32_t, const void *); /* logger */
 extern int gatt_req_send(void *, void *, void *, void *, uint32_t, uint32_t);
-extern int memcmp(const void *, const void *, uint32_t); /* memcmp */
+extern int memcmp(const void *, const void *, size_t); /* memcmp */
 extern int bt_gatt_write(void *, void *); /* bt_gatt_write */
 
 struct write_params_recovered {
