+++ recon/symbolized/app/adv_is_directed.c
@@ -1,4 +1,5 @@
 #include "g1_app_symbols.h"
+#include <stddef.h>
 /* readable reconstruction; identity: FUN_00055698 @ 0x00055698
  * public-name: adv_is_directed
  * durable-map: recon/catalogs/function_names_app.json
@@ -9,5 +10,5 @@
  * Raw backmap: FUN_00055698@0x00055698. */
 
 #include <stdint.h>
-extern int memcmp(const void*,const void*,uint32_t);
+extern int memcmp(const void *, const void *, size_t);
 int adv_is_directed(unsigned long advertisement) {return memcmp((const uint8_t*)advertisement+9u,(const void*)((unsigned long)&rodata_f2b3a) /*=0xf2b3a*/,7u)!=0;}
